+++ src/pyday/DataReader/DataReader.py
@@ -3,28 +3,21 @@ import sys
 import pandas as pd
 
 from ..basic import config
+from ..basic.BasisDay import BasisDay
 
-class DataReader:
+class DataReader(BasisDay):
     def __init__(self, inFile=None):
+        super().__init__("reader")
         self.fileName = inFile
-        self.inPath = os.path.join( config.worKdir, "pydayData/Reader" )
-        self.toPath = os.path.join( config.worKdir, "pydayDist/Reader" )
         self.df = None
-
-        if not "ipykernel" in sys.modules:
-            self.loadDir()
-            
+        
         if isinstance(inFile, str):
             self.inFile(inFile)
         else:
             self.df = pd.DataFrame(inFile)
-
-    def loadDir(self):
-        for path in [self.inPath, self.toPath]:
-            if not os.path.exists(path):
-                os.makedirs(path, exist_ok=True)
     
     def inFile(self, inFile):
+        self.loadDir( [self.inPath] )
         inFileFormat = os.path.splitext(inFile)[-1][1:].lower()
         inFile = os.path.join(self.inPath, inFile)
         if os.path.exists(inFile):
@@ -39,21 +32,14 @@ class DataReader:
         else:
             print( f"Did Not Have File: {inFile}" )
 
-    def setPath(self, inPath):
-        if not os.path.exists(inPath):
-            os.mkdir(inPath)
-        self.inPath = inPath
-
-    def getToPath(self):
-        return self.toPath
-
     def toFile(self, toFile):
+        self.loadDir( [self.toPath] )
         toFileFormat = os.path.splitext(toFile)[-1][1:].lower()
         toFile = os.path.splitext(toFile)[0]
-        out = f"{self.getToPath()}/{toFile}"
+        out = f"{self.toPath}/{toFile}"
         if toFileFormat == "all":
             self.df.to_csv( f"{out}.csv", index=False)
-            self.df.to_json( f"{out}.json" )
+            self.df.to_json( f"{out}.json")
             self.df.to_excel( f"{out}.xlsx" )
         elif toFileFormat == "csv":
             self.df.to_csv( f"{out}.csv", index=False)
@@ -63,8 +49,3 @@ class DataReader:
             self.df.to_excel( f"{out}.xlsx", index=False)
         else:
             print( "output: Does not support" )
-
-    def setToPath(self, toPath):
-        if not os.path.exists(toPath):
-            os.mkdir(toPath)
-        self.toPath = toPath