+++ src/pyday/basic/config.py
@@ -3,6 +3,7 @@ import os
 # 獲取 目前腳本pyday所在目錄的父目錄的絕對路徑
 basedir = os.path.dirname(os.path.abspath(os.path.dirname(__file__)))
 worKdir = os.getcwd()
+
 # 字體的絕對路徑
 font_path = os.path.join(basedir, "basic/font")
 Noto_Sans_TC = os.path.join(font_path, "Noto_Sans_TC")